DjangoRestApi/requirements.txt
asgiref==3.2.7
astroid==2.4.2
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
coverage==5.1
Django==3.0.7
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
django-storages==1.9.1
django-phone-field==1.8.1
google-api-core==1.20.1
google-auth==1.17.2
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
#Pillow==7.1.2
protobuf==3.12.2
#psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
social-auth-app-django